[bootstrap] Refactor bootstrap script
This separates the core "bootstrapping" scripts I need to get a basic
environment up and running and subsequent "setup" scripts that can be
run later as smaller, more isolated scripts.

The goal here is to make most of the setup more easily testable, and to
make it easier to recover if something goes wrong with an given
component.