[bootstrap] Fix onepassword session handling
The output of `op signin` is an `export ...` line that needs to be
sourced into other scripts.  This means that just running it in the
1password setup script isn't enough to propagate the session to other
setup scripts.  However, by pushing that output to a temp file, we can
source that file in the subsequent setup scripts and get a usable
session.